Add test that loads from Config::Merge
I had a weird situation where this code would work correctly
on my Mac and Ubuntu 11, but not on my Ubuntu 10 installation.
All other tests worked correctly and when I fed the exact same
hash as created by Config::Merge using a static declaration,
it also worked correctly. When I used Config::Merge, the sha1
that was generated for tree objects was erroneous and caused
the git bare repository to be corrupt.

I resolved the problem by re-installing my Ubuntu 10. I saved
this test for the event that I run into this problem again.